Entrega final

+++ src/main/java/tudai/prog3/algoritmo/Algoritmo.java
@@ -2,25 +2,23 @@ package tudai.prog3.algoritmo;
 
 import tudai.prog3.colecciones.Estado;
 
-
 public abstract class Algoritmo {
 
-    protected String nombre;
-    protected int iteraciones;
+	protected String nombre;
+	protected long iteraciones;
 
-    public Algoritmo(String name) {
-        this.nombre = name;
-    }
+	public Algoritmo(String name) {
+		this.nombre = name;
+	}
 
-    public abstract Estado hallarRedDeMenorLongitud(Estado ei);
+	public abstract Estado hallarRedDeMenorLongitud(Estado ei);
 
-    public String getNombre() { return this.nombre; }
+	public String getNombre() {
+		return this.nombre;
+	}
 
-    public int getIteraciones() {
-        return this.iteraciones;
-    }
-
-
-
+	public long getIteraciones() {
+		return this.iteraciones;
+	}
 
 }
